// src/app/manga/services/__generated__/MangaFragment.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL fragment: MangaFragment
// ====================================================

export interface MangaFragment {
    __typename: "Manga";
    id: any;
    isLikedByViewer: boolean;
    likesCount: number;
}